% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dc_causality.R
\name{dc_causality}
\alias{dc_causality}
\alias{dc_causality_sub}
\title{Extract Causality Discourse Connectors in Context}
\usage{
dc_causality(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"),
  name = "causality", ...)

dc_causality_sub(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"),
  name = "causality_sub", ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{grouping.var}{The grouping variables.  Also takes a single
grouping variable or a list of 1 or more grouping variables.}

\item{n.before}{The number of rows before the indexed occurrence.}

\item{tot}{logical.  If \code{TRUE} condenses sub-units (e.g., sentences)
into turns of talk for that \code{grouping.var}.}

\item{n.after}{The number of rows after the indexed occurrence.}

\item{ord.inds}{logical.  If \code{TRUE} inds is ordered least to greatest.}

\item{markup}{A character vector of length two indicating the left (element
1) and right (element 2) boundary markers to use to highlight the causality
discourse connectors.  Use \code{c("", "")} to not mark the causality discourse
markers.}

\item{name}{A string indicating the name to search for within the internal
data sets, typically the function's name.  Generally, for internal use.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{termco}}.}
}
\value{
\code{dc_causality} - Returns returns a list of 2:
\item{counts}{A \code{\link[qdap]{termco}} object of causality discourse connector counts.}
\item{causality}{A \code{\link[qdap]{trans_context}} object of causality discourse connectors in context.}

\code{dc_causality_sub} - Returns returns a list of 3:
\item{counts}{A \code{\link[qdap]{termco}} object of causality discourse connector counts.}
\item{continuation}{A \code{\link[qdap]{trans_context}} object of continuation causality discourse connectors in context.}
\item{elaboration}{A \code{\link[qdap]{trans_context}} object of elaboration causality discourse connectors in context.}
}
\description{
\code{dc_causality} - Extract causality discourse connectors in context.

\code{dc_causality_sub} - Extract causality discourse connectors in context typified as
continuation or elaboration.
}
\examples{
with(pres_debates2012[1:200, ], dc_causality(dialogue, person))

out <- with(pres_debates2012[1:200, ], dc_causality_sub(dialogue, person))
out[1]
out[2]
out[3]
plot(out)

## Save externally use .doc or .txt
## print(out[[2]], file="elaboration_causality.doc")
}
\references{
Alemany, L. A. (2005). Representing discourse for automatic text summarization via shallow NLP techniques (Unpublished doctoral dissertation). Universitat de Barcelona, Barcelona.

\url{http://russell.famaf.unc.edu.ar/~laura/shallowdisc4summ/discmar}
}
\seealso{
\code{\link[qdap]{termco}},
\code{\link[qdap]{trans_context}}
}
\keyword{causality}

