% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dc_equality.R
\name{dc_equality}
\alias{dc_equality}
\alias{dc_equality_sub}
\title{Extract Equality Discourse Connectors in Context}
\usage{
dc_equality(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"),
  name = "equality", ...)

dc_equality_sub(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"),
  name = "equality_sub", ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{grouping.var}{The grouping variables.  Also takes a single
grouping variable or a list of 1 or more grouping variables.}

\item{n.before}{The number of rows before the indexed occurrence.}

\item{tot}{logical.  If \code{TRUE} condenses sub-units (e.g., sentences)
into turns of talk for that \code{grouping.var}.}

\item{n.after}{The number of rows after the indexed occurrence.}

\item{ord.inds}{logical.  If \code{TRUE} inds is ordered least to greatest.}

\item{markup}{A character vector of length two indicating the left (element
1) and right (element 2) boundary markers to use to highlight the equality
discourse connectors.  Use \code{c("", "")} to not mark the equality discourse
markers.}

\item{name}{A string indicating the name to search for within the internal
data sets, typically the function's name.  Generally, for internal use.}

\item{inds}{A list of integer indices to print context for.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{termco}}.}
}
\value{
\code{dc_equality} - Returns returns a list of 2:
\item{counts}{A \code{\link[qdap]{termco}} object of equality discourse connector counts.}
\item{equality}{A \code{\link[qdap]{trans_context}} object of equality discourse connectors in context.}

\code{dc_equality_sub} - Returns returns a list of 4:
\item{counts}{A \code{\link[qdap]{termco}} object of equality discourse connector counts.}
\item{none}{A \code{\link[qdap]{trans_context}} object of equality discourse connectors in context that are not \emph{continuation} or \emph{elaboration}.}
\item{continuation}{A \code{\link[qdap]{trans_context}} object of continuation equality discourse connectors in context.}
\item{elaboration}{A \code{\link[qdap]{trans_context}} object of elaboration equality discourse connectors in context.}
}
\description{
\code{dc_equality} - Extract equality discourse connectors in context.

\code{dc_equality_sub} - Extract equality discourse connectors in context typified as
continuation or elaboration.
}
\examples{
with(pres_debates2012[1:200, ], dc_equality(dialogue, person))

out <- with(pres_debates2012[1:200, ], dc_equality_sub(dialogue, person))
out[1]
out[2]
out[3]
out[4]
plot(out)

## Save externally use .doc or .txt
## print(out[[3]], file="elaboration_equality.doc")
}
\references{
Alemany, L. A. (2005). Representing discourse for automatic text summarization via shallow NLP techniques (Unpublished doctoral dissertation). Universitat de Barcelona, Barcelona.

\url{http://russell.famaf.unc.edu.ar/~laura/shallowdisc4summ/discmar}
}
\seealso{
\code{\link[qdap]{termco}},
\code{\link[qdap]{trans_context}}
}
\keyword{equality}

