% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dc_oh.R
\name{dc_oh}
\alias{dc_oh}
\title{Extract Oh Discourse Connectors in Context}
\usage{
dc_oh(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"),
  name = "oh", ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{grouping.var}{The grouping variables.  Also takes a single
grouping variable or a list of 1 or more grouping variables.}

\item{n.before}{The number of rows before the indexed occurrence.}

\item{tot}{logical.  If \code{TRUE} condenses sub-units (e.g., sentences)
into turns of talk for that \code{grouping.var}.}

\item{n.after}{The number of rows after the indexed occurrence.}

\item{ord.inds}{logical.  If \code{TRUE} inds is ordered least to greatest.}

\item{markup}{A character vector of length two indicating the left (element
1) and right (element 2) boundary markers to use to highlight the oh
discourse connectors.  Use \code{c("", "")} to not mark the oh discourse
markers.}

\item{name}{A string indicating the name to search for within the internal
data sets, typically the function's name.  Generally, for internal use.}

\item{inds}{A list of integer indices to print context for.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{termco}}.}
}
\value{
Returns returns a list of 2:
\item{counts}{A \code{\link[qdap]{termco}} object of oh discourse marker counts.}
\item{oh}{A \code{\link[qdap]{trans_context}} object of oh discourse connectors in context.}
}
\description{
Extract oh discourse connectors in context.
}
\examples{
out <- with(pres_debates2012[1:1000, ], dc_oh(dialogue, person))
out[1]
out[2]
plot(out)

## Save externally use .doc or .txt
## print(out[[2]], file="oh.doc")
}
\references{
Schiffrin, D. (1987). \emph{Discourse markers}.  (pp. 73-101). London: Cambridge University Press.
}
\seealso{
\code{\link[qdap]{termco}},
\code{\link[qdap]{trans_context}}
}
\keyword{oh}

