% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discon-package.R
\docType{package}
\name{discon}
\alias{discon}
\alias{discon-package}
\alias{package-discon}
\title{discon: Tools for analysing discourse markers in text}
\description{
\pkg{discon} is a collection of R tools for analysing discourse connectors in
text.  Discourse connectors are cohesive devices that can be used to help
identify themes within a text.  This package provides computational means of
extracting various elements from the text that contain discourse connectors.
This can assist in qualitative analysis of discourse by identifying categories
that may aide analysis (using the computer for efficiency and data coverage)
towards generating themes.
}
\details{
``Discourse connectors are devices used to bridge between turns (in speech)
and sentences, indicating the logical relations among the parts of a the
logical relations among the parts of a framework for the listener/reader.
There are two major classes of discourse connectors: discourse markers and
linking adverbials. Discourse markers -- forms like \emph{ok},
\emph{well}, and \emph{now} -- are restricted primarily to spoken
discourse. These forms have distinct discourse functions, but it is difficult
to identify the specific meaning of the word itself. In contrast, linking
adverbials -- forms like \emph{however}, \emph{thus},
\emph{therefore}, \emph{for example} (\emph{e.g.}), and \emph{that is}
(\emph{i.e.}) -- are found in both spoken and written registers, and
they have greater inherent meaning than discourse markers.'' (Beiber, 2006,
p. 66)

Please see the following resources for additional information:

\itemize{
\item Alemany, L. A. (2005). \href{http://www.cs.famaf.unc.edu.ar/~laura/shallowdisc4summ/tesi_electronica.pdf}{Representing discourse for automatic text summarization via shallow NLP techniques} (Unpublished doctoral dissertation). Universitat de Barcelona, Barcelona.
\item Biber, D. (2006). \emph{University language: A corpus-based study of spoken and written registers}. Amsterdam: John Benjamins Publishing Company.
\item Halliday, M. A. K. & Hasan, R. (1976). \emph{Cohesion in English}. English Language Series. London: Longman.
\item Kalajahi, S. A. R., Abdullah, A. N., Mukundan, J., & Tannacito, D. J. (2012). \href{http://goo.gl/eS0OwV}{Discourse connectors: An overview of the history, definition and classification of the term}. \emph{World Applied Sciences Journal, 19}(11), 1659-1673.
\item Linguistic Data Consortium. (2004) \href{http://www.itl.nist.gov/iad/mig/tests/rt/2004-fall/docs/SimpleMDE_V6.2-draft.pdf}{\emph{Simple metadata annotation specification (MDE)}}. University of Pennsylvania.  Retrieved from \url{http://www.itl.nist.gov/iad/mig/tests/rt/2004-fall/docs/SimpleMDE_V6.2-draft.pdf}
\item Ryan, G. W. & Bernard, H. R. (2003). \href{http://goo.gl/KdYxB8}{Techniques to identify themes}. \emph{Field Methods. 15}(1), 85-109. doi: \href{http://fmx.sagepub.com/content/15/1/85}{10.1177/1525822X02239569}
\item Schiffrin, D. (1987). \emph{Discourse markers}. London: Cambridge University Press.
\item Schiffrin, D. (2001). \href{http://goo.gl/KbwG9I}{Discourse markers: Language, meaning, and context}. In D. Schiffrin, D., D. Tannen, & H. E. Hamilton, (Eds.) \emph{The handbook of discourse analysis} (pp. 54-75). Malden, MA: Blackwell Publishing.
\item Ward, N. (2000, October). \href{http://www.aclweb.org/anthology/W00-1004}{Issues in the Transcription of English Conversational Grunts}. \emph{Proceedings of the 1st SIGdial Workshop on Discourse and Dialogue} (pp. 29-35). Hong Kong: Association for Computational Linguistics.  Retrieved from \url{http://www.aclweb.org/anthology/W00-1004}
}
}

