% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discourse_connector.R
\name{discourse_connector}
\alias{discourse_connector}
\title{Extract Discourse Connectors in Context}
\usage{
discourse_connector(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"),
  name = NULL, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{grouping.var}{The grouping variables.  Also takes a single
grouping variable or a list of 1 or more grouping variables.}

\item{n.before}{The number of rows before the indexed occurrence.}

\item{tot}{logical.  If \code{TRUE} condenses sub-units (e.g., sentences)
into turns of talk for that \code{grouping.var}.}

\item{n.after}{The number of rows after the indexed occurrence.}

\item{ord.inds}{logical.  If \code{TRUE} inds is ordered least to greatest.}

\item{markup}{A character vector of length two indicating the left (element
1) and right (element 2) boundary markers to use to highlight the
discourse connectors.  Use \code{c("", "")} to not mark the discourse
connectors.}

\item{name}{A string indicating the name to search for within the internal
data sets, typically the function's name.  Generally, for internal use.}

\item{inds}{A list of integer indices to print context for.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{termco}}.}
}
\value{
Returns returns a list of 2-3:
\item{counts}{A \code{\link[qdap]{termco}} object of discourse connector counts.}
\item{Context 1}{A \code{\link[qdap]{trans_context}} object of the discourse connectors in context.  Note the name of this object is supplied by \code{names} element one.}
\item{Context 2...n}{An optional (not returned if \code{regex} is of length one) \code{\link[qdap]{trans_context}} object of the discourse connectors in context.  Note the name of this (these) object(s) is supplied by \code{names} element 2...n.}
}
\description{
Extract discourse connectors in context.  This is the flexible default template
for modular use in specific discourse connector functions.
}
\examples{
## Marker with one type (just: "I")
out1 <- with(pres_debates2012[1:200, ], discourse_connector(dialogue, person,
    names = c("I"),
    regex = "\\\\bI('[a-z]+)*\\\\b",
    terms = list(I = c(" I ", " I'"))
))

out1[[1]]
out1[[2]]
plot(out1)

## Marker with two types (both: "I" & "you")
out2 <- with(pres_debates2012[1:200, ], discourse_connector(dialogue, person,
    names = c("I", "you"),
    regex =  list(
        I = "I('[a-z]+)*\\\\b",
        you = "(\\\\b[Yy]ou('[a-z]+)*\\\\b)"
    ),
    terms = list(
        I = c(" I ", " I'"),
        you = c(" you ", " you'")
    )
))
out2[[1]]
out2[[2]]
out2[[3]]

## Save externally use .doc or .txt
## print(out2[[2]], file="you_I.doc")
}
\references{
Kalajahi, S. A. R., Abdullah, A. N., Mukundan, J., & Tannacito, D. J. (2012) \href{http://goo.gl/eS0OwV}{Discourse connectors: An overview of the history, definition and classification of the term}. \emph{World Applied Sciences Journal, 19}(11), 1659-1673.
}
\seealso{
\code{\link[qdap]{termco}},
\code{\link[qdap]{trans_context}}
}
\keyword{connector}
\keyword{discourse}

