% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.first_in_set.R
\name{is.first_in_set}
\alias{is.first_in_set}
\alias{is.last_in_set}
\title{Check if Element Is First in a Set}
\usage{
is.first_in_set(x)

is.last_in_set(x)
}
\arguments{
\item{x}{A vector of elements.}
}
\value{
Returns a logical vector equal in length to x.
}
\description{
\code{is.first_in_set} - Logical check of each element in a vector to
determine if an element is first in set.

\code{is.last_in_set} - The (semantical) "inverse" to \code{is.first_in_set}.
}
\examples{
set.seed(3)
(x <- sample(1:3, 9, TRUE))
is.first_in_set(x)
is.last_in_set(x)
}

