% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.isolate.R
\name{is.isolate}
\alias{is.isolate}
\alias{is.repeat}
\title{Check if Element Is Isolated}
\usage{
is.isolate(x)

is.repeat(x)
}
\arguments{
\item{x}{A vector of elements.}
}
\value{
Returns a logical vector equal in length to x.
}
\description{
\code{is.isolate} - Logical check of each element in a vector to determine if
an element is isolated, that is not repeated.

\code{is.repeat} - The (semantical) "inverse" to \code{is.isolate}.
}
\examples{
set.seed(3)
(x <- sample(1:3, 9, TRUE))
is.isolate(x)
is.repeat(x)
}

