% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.within_n_preceding_words.R
\name{is.within_n_preceding_words}
\alias{is.within_n_preceding_words}
\title{Check if Element Is Within N Preceding Words}
\usage{
is.within_n_preceding_words(x, regex.term, n = 0)
}
\arguments{
\item{x}{A vector of elements.}

\item{regex.term}{A regular expression that represents a word or term.}

\item{n}{A max number of words (defined as \code{([a-zA-Z-]+,*\\s+)}) that
\code{regex.term} must occur before.}
}
\value{
Returns a logical vector equal in length to x.
}
\description{
Logical check of each element in a vector to determine if
an element occurs within n preceding words.
}
\examples{
x <- c(
    "I like dogs and cats",
    "I do not like robots though",
    "Like I care."
)

is.within_n_preceding_words(x, "I", 0)
is.within_n_preceding_words(x, "I", 1)
is.within_n_preceding_words(x, "like", 4)
is.within_n_preceding_words(x, "[Ll]ike", 4)
}

