% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kwic.R
\name{kwic}
\alias{kwic}
\title{Extract Key Words in Context}
\usage{
kwic(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"), top = 15,
  at.least = 5, extend = TRUE, ignore.case = FALSE, left = "\\\\b",
  right = "", names = c("KeyWords"), elim.old = FALSE,
  stopwords = c("going", qdapDictionaries::contractions[[1]],
  qdapDictionaries::Top200Words), ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{grouping.var}{The grouping variables.  Also takes a single
grouping variable or a list of 1 or more grouping variables.}

\item{n.before}{The number of rows before the indexed occurrence.}

\item{tot}{logical.  If \code{TRUE} condenses sub-units (e.g., sentences)
into turns of talk for that \code{grouping.var}.}

\item{n.after}{The number of rows after the indexed occurrence.}

\item{ord.inds}{logical.  If \code{TRUE} inds is ordered least to greatest.}

\item{markup}{A character vector of length two indicating the left (element
1) and right (element 2) boundary markers to use to highlight the
key words.  Use \code{c("", "")} to not mark the discourse
connectors.}

\item{top}{Top number of terms to show.}

\item{at.least}{An integer indicating at least how many letters a word
must be to be included in the output.}

\item{extend}{logical.  If \code{TRUE} the \code{top} argument is extended to
any word that has the same frequency as the \code{top} word.}

\item{ignore.case}{logical.  If \code{TRUE} leading character's case is
ignored.}

\item{left}{The left boundary (regular expression) to put on the left side of
words.  Default is
\href{http://www.regular-expressions.info/wordboundaries.html}{word boundary}
(i.e., \code{\\\\b}) that cuts off at non-word characters.}

\item{right}{The right boundary (regular expression) to put on the right side
of words.  Default is no boundary and will locate words within words.  For
example the expression \code{"\\\\blast"} will find "last" and "lastly" but not
"blast".  To capture precisely key words use \code{right = "\\\\b"} or
\code{right = left}.}

\item{names}{A string indicating the name of the key words in the collective
group name for the key words for use in the \code{\link[qdap]{termco}} and
generic \code{plot} (\code{\link[qdap]{dispersion_plot}}) functions.}

\item{elim.old}{logical.  If \code{TRUE} eliminates the columns that are
combined together by the named match.list.}

\item{stopwords}{A character vector of words to remove from the text.}

\item{inds}{A list of integer indices to print context for.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{termco}}.}
}
\value{
Returns returns a list of 2-3:
\item{counts}{A \code{\link[qdap]{termco}} object of discourse connector counts.}
\item{Context 1}{A \code{\link[qdap]{trans_context}} object of the key words in context.  Note the name of this object is supplied by \code{names} element one.}
\item{Context 2...n}{An optional (not returned if \code{regex} is of length one) \code{\link[qdap]{trans_context}} object of the key words in context.  Note the name of this (these) object(s) is supplied by \code{names} element 2...n.}
}
\description{
Extract key words in context (kwic) utilizing \code{\link[qdap]{freq_terms}}.
}
\details{
While \code{kwic} does not analyze discourse connectors, it is useful
in identifying themes (and distribution across time) and builds upon the
modular use of functions in the pkg{discon} package.
}
\examples{
out <- with(pres_debates2012, kwic(dialogue, list(time, person)))
plot(out[[1]])
plot(out)
plot(out, rm.vars = pres_debates2012[["time"]], total.color = NULL)
}
\references{
Ryan, G. W. & Bernard, H. R. (2003). \href{http://goo.gl/KdYxB8}{Techniques to identify themes}. \emph{Field Methods. 15}(1), 85-109. doi: \href{http://fmx.sagepub.com/content/15/1/85}{10.1177/1525822X02239569}
}
\seealso{
\code{\link[qdap]{termco}},
\code{\link[qdap]{trans_context}},
\code{\link[qdap]{freq_terms}}
}
\keyword{frequent}
\keyword{key}
\keyword{kwic}
\keyword{words}

