% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discourse_connector.R
\name{plot.discourse_connector}
\alias{plot.discourse_connector}
\title{Plots a discourse_connector Object}
\usage{
\method{plot}{discourse_connector}(x, grouping.var = NULL,
  unlist.terms = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{The discourse_connector object.}

\item{grouping.var}{The grouping variables.  Take from \code{x} unless
supplied by the user.}

\item{unlist.terms}{logical.  If \code{TRUE} the terms will be separated rather than grouped as a single (sub)marker.}

\item{plot}{logical.  If \code{TRUE} the plot will automatically plot. The
user may wish to set to \code{FALSE} for use in \pkg{knitr}, \pkg{sweave},
etc. to add additional plot layers.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{dispersion_plot}}.}
}
\description{
Plots a discourse_connector object.
}

