% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discourse_connector.R
\name{print.discourse_connector}
\alias{print.discourse_connector}
\title{Prints an discourse_connector Object}
\usage{
\method{print}{discourse_connector}(x, ...)
}
\arguments{
\item{x}{The discourse_connector object.}

\item{\ldots}{ignored.}
}
\description{
Prints an discourse_connector object.
}

