% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dc_oh.R
\name{dc_oh}
\alias{dc_oh}
\alias{dc_oh_begin}
\title{Extract Oh Discourse Connectors in Context}
\usage{
dc_oh(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"),
  name = "oh", ...)

dc_oh_begin(text.var, grouping.var, n.before = 1, tot = FALSE,
  n.after = n.before, ord.inds = TRUE, markup = c("<<", ">>"),
  name = "oh", fun1, fun2 = NULL, ...)
}
\arguments{
\item{text.var}{The text variable.}

\item{grouping.var}{The grouping variables.  Also takes a single
grouping variable or a list of 1 or more grouping variables.}

\item{n.before}{The number of rows before the indexed occurrence.}

\item{tot}{logical.  If \code{TRUE} condenses sub-units (e.g., sentences)
into turns of talk for that \code{grouping.var}.}

\item{n.after}{The number of rows after the indexed occurrence.}

\item{ord.inds}{logical.  If \code{TRUE} inds is ordered least to greatest.}

\item{markup}{A character vector of length two indicating the left (element
1) and right (element 2) boundary markers to use to highlight the oh
discourse connectors.  Use \code{c("", "")} to not mark the oh discourse
markers.}

\item{name}{A string indicating the name to search for within the internal
data sets, typically the function's name.  Generally, for internal use.}

\item{fun1}{A function that checks the text variable and returns a logical
vector.  This allows for additional restrictions to be places upon the text
beyond the limited (non-regex) capabilities of \code{\link[qdap]{termco}} and
\code{\link[qdap]{trans_context}}.  The function in \code{dc_oh_begin}
ensures that each has no more than n words before the "oh".  The defaut number
of words is 0 or less.  This can be changed by supplying an argument to
\code{n} via \code{control}.  For example to set the number of words to 2
use: \code{control = list(n = 2))}.}

\item{fun2}{A function that checks the grouping variable and returns a logical
vector.  This allows for additional restrictions to be placed upon the
grouping variables that can't be addressed by \code{\link[qdap]{termco}} and
\code{\link[qdap]{trans_context}}.  This argument in \code{dc_oh_begin} is
ignored.}

\item{inds}{A list of integer indices to print context for.}

\item{\ldots}{Other arguments passed to \code{\link[qdap]{termco}}.}
}
\value{
Returns returns a list of 2:
\item{counts}{A \code{\link[qdap]{termco}} object of oh discourse marker counts.}
\item{oh}{A \code{\link[qdap]{trans_context}} object of oh discourse connectors in context.}
}
\description{
\code{dc_oh} - Extract oh discourse connectors (a marker Schiffrin (1987)
terms a marker of "information management") in context.

\code{dc_oh_begin} - An extension of \code{dc_oh} that requires the "oh" to
come at the begining of the word.
}
\examples{
out <- with(pres_debates2012[1:1000, ], dc_oh(dialogue, person))
out[1]
out[2]
plot(out)

## Save externally use .doc or .txt
## print(out[[2]], file="oh.doc")
}
\references{
Schiffrin, D. (1987). \emph{Discourse markers}.  (pp. 73-101). London: Cambridge University Press.
}
\seealso{
\code{\link[qdap]{termco}},
\code{\link[qdap]{trans_context}}
}
\keyword{oh}

