% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discon-package.R
\docType{package}
\name{discon}
\alias{discon}
\alias{discon-package}
\alias{package-discon}
\title{discon: Tools for analysing discourse markers in text}
\description{
\pkg{discon} is a collection of R tools for analysing discourse connectors in
text.  Discourse connectors are cohesive devices that can be used to help
identify themes within a text.  This package provides computational means of
extracting various elements from the text that contain discourse connectors.
This can aid in qualitative analysis of discourse by identifying categories
that may aide analysis (using the computer for efficiency and data coverage)
towards generating themes.
}
\details{
Please see the following resources for additional information:

\itemize{
\item Alemany, L. A. (2005). \href{http://www.cs.famaf.unc.edu.ar/~laura/shallowdisc4summ/tesi_electronica.pdf}{Representing discourse for automatic text summarization via shallow NLP techniques} (Unpublished doctoral dissertation). Universitat de Barcelona, Barcelona.
\item Halliday, M. A. K. & Hasan, R. (1976). \emph{Cohesion in English}. English Language Series. London: Longman.
\item Kalajahi, S. A. R., Abdullah, A. N., Mukundan, J., & Tannacito, D. J. (2012) \href{http://goo.gl/eS0OwV}{Discourse connectors: An overview of the history, definition and classification of the term}. \emph{World Applied Sciences Journal, 19}(11), 1659-1673.
\item Ryan, G. W. & Bernard, H. R. (2003). \href{http://goo.gl/KdYxB8}{Techniques to identify themes}. \emph{Field Methods. 15}(1), 85-109. doi: \href{http://fmx.sagepub.com/content/15/1/85}{10.1177/1525822X02239569}
\item Schiffrin, D. (1987). \emph{Discourse markers}. London: Cambridge University Press.
\item Schiffrin, D. (2001). \href{http://goo.gl/KbwG9I}{Discourse markers: Language, meaning, and context}. In D. Schiffrin, D., D. Tannen, & H. E. Hamilton, (Eds.) \emph{The handbook of discourse analysis} (pp. 54-75). Malden, MA: Blackwell Publishing.
}
}

